chore(team): split team workspace structure
